+++ docs/pages/demo.js
@@ -3,7 +3,7 @@ import dynamic from 'next/dynamic'
 
 const DynamicComponentWithNoSSR = dynamic(
   () => import('../src/components/screens/Demo'),
-  { ssr: false }
+  { ssr: true }
 )
 
 export default class Demo extends Component {
