+++ src/CSVReader.js
@@ -193,6 +193,7 @@ export default class CSVReaderRewrite extends Component {
   }
 
   render() {
+    
     const {
       style,
       noClick,
